Resolve ResourceNotFound error on transcript generation
When trying to write out the transcript file a `ResourceNotFound` exception was being raised. This is because the file path that it was attempting to write to was not present. The `delegate_fs` function call resulted in the root path being one level up from the course output directory. In addition, prepending `course` to the static_dir was erroneous, resulting in a double mismatch of the target directory. Using the `resource_fs` directly and removing the `course` path prefix results in a correct target for the transcript file to be written.